Assorted updates to fix dependencies and Typescript
This includes some new packages which will be useful to a later
commit. Since getting Ember and Typescript working required a
significant amount of trial and error, I did not attempt to
separate packages required for the later commit from packages
required for the initial setup of Ember and Typescript.