Create utility classes to represent a single die and a group of dice
A single die and a group of identical dice are the basic building
blocks of attacks and damage in D&D 5e and similar systems. This
defines classes for each of these and adds tests for them.
Isolating the "die" concept into a separate class makes mocking
behavior significantly easier for later tests.

This deletes the auto-created tests/helpers/index.js file because
it did not work well with TypeScript, whereas the default
ember-qunit file worked smoothly.