Introduce parser for strings combining dice groups and numbers
This introduces a parser for strings consisting of groups of dice
and numbers added or subtracted from one another. Attack and
damage information in systems like D&D 5e is represented in this
format (eg "1d20 + 5 - 1d6" for an attack roll affected by Synaptic
Static or "1d6 + 3 + 3d8" for a paladin's attack and smite). Parsing
these strings into a series of dice groups and a constant will allow
users to type in attack details in a familiar format.