Add utility class representing a damage type
This class is useful for calculating the damage of a particular
type inflicted by an attack. Since targets may resist some damage
types and be vulnerable to others, handling each type as a separate
class simplifies damage calculations for attacks which work with
multiple damage types.

Although there will be a section of the input form dedicated
to each damage type, this class interacts minimally with the data
in the form and will be created as part of the overall attack
creation, so it is not currently linked to an Ember component.