Create class to represent an attack with its details
This creates a class to represent an attack (with a to-hit
modifier and one or more associated damage types) and allow
attacks with varying target AC and advantage/disadvantage
conditions.

It also fixes a small bug in the parser for dice strings,
which did not handle single digits with a sign correctly.