Fully remove ci.yml configuration for Github actions
Even with the jobs commented out, Github still attempted to run CI
as long as this file is present. As described in a previous commit,
this job was auto-generated and did not function, and fixing it
is a low-priority task as long as only one developer is contributing
to this repository.