Add a form to represent repeated attacks in a system like D&D 5e
This adds a form which takes the details for repeated attacks
and uses them to simulate such attacks and display the result
in a text pane.

The form has a few bugs (for instance, the advantage/disadvantage
radio buttons do not un-set the other variables when a new button
is selected) and needs significant work to make the display more
attractive (likely adding Bootstrap) and flexible (isolating
damage into its own component and then supporting multiple damage
types). However, these improvements will be handled in later
commits to make the changes easier to understand.