Factor out damage type component and support different damage types
This factors out a component to represent a single damage type,
including the name of that type, the associated damage string,
and whether the target is resistant or vulnerable to that damage
type. It also updates this component to include a dropdown list
of possible damage types.

This adds the ember-truth-helpers plugin to support "eq" comparison
for strings in the Handlebars files, allowing the dropdown to
be initialized with a damage type selected.