Reduce repeated code in advantage radio button
This introduces a loop over possible advantage states to reduce
repeated code in the radio button component. Unfortunately,
Ember's radio button input does not support a "checked" attribute
and the ember-radio-button and ember-radio-buttons plugins are
several years out of date. This means that it does not appear
possible to have the buttons initialized with one of the values
already selected (if we used checkboxes, this would be possible,
but this would lose the constraint that only one button can be
selected).