Decapitalize damage types' names
This removes capitalization from damage types, since these are
typically written in lower-case. It also rewrites several tests
to stop using acid damage. Since "acid" is the first entry in the
select menu for damage types (which is organized alphabetically),
if the "select(dropdown, option)" command in a test fails (as when
an invalid type is submitted), "acid" is set as the default result.
This obscured some test failures caused by the "data-test-*" syntax
not working for the input option.