Update regular expression for dice strings to be case insensitive
Although the DiceStringParser class already handles dice strings
in a case insensitive way, the validity check for the toHit and
damage fields is case-sensitive. Since HTML's pattern attribute
cannot be made case-insensitive, this updates the regular
expression to allow "1d6" or "1D6" equally. This makes the form
easier to use and avoids invalidating the form for terms which
the backend can actually handle.