Use CSS classes rather than undefined-element behavior
This updates the custom bullet points for lists to use classes,
rather than relying on the fact that an undefined element is
handled as an empty element and then formatted as specified in the
file. It also adjusts the numbers in the test case to better match
the expected values from the described dice rolls.