Display the number of successful hits in the attack summary
This displays how many hits were successful as well as the
total damage in the summary of an attack. This is useful for
situations where additional action must be taken on a successful
hit, such as a constitution save against poison or lycanthropy,
since it makes it easy to count how many saves should be made.