Add log of previous repeated attacks
This adds a log of previous repeated attacks, as well as adjusting
the format of displayed data to reduce redundant information
(for instance, damage details are removed since they are usually
easily visible in the attack log). The log automatically scrolls
once it exceeds the size of the viewport.

The log is currently not stored anywhere, so is erased upon page
refresh. Later work will save it to local or session storage
and add an option for manually clearing the log.