Display doubled dice groups upon critical hits
This changes the attack display so that when a critical hit is
rolled, the damage dice are printed with double the usual number
of dice. This reflects that when a critical hit is rolled, twice
the number of dice are rolled (although the modifier is not doubled).
For instance, "2d6 + 4" damage prints as "4d6 + 4" on a critical hit.

This also standardizes the printing for damage dice and to-hit
modifiers so that there are spaces between the terms and the dice
groups are printed before the modifier. This is potentially
slightly confusing, since it does not exactly match the text entered
into the form; however, it simplifies doubling the dice groups and
makes the display more consistent.