Improve client-side validation and add server-side validation
This improves the server-side checks for missing or invalid fields
in the repeated attack and damage classes, and rejects attacks if
the attack or damage are not constructed properly. It also marks
all fields as required in the form, so that users cannot manually
delete the values and submit with no error indications.