Display dice rolls in hovertext (#176)
This adds hovertext to both attack and damage rolls which displays the actual
values rolled on the dice. This is useful for various edge cases (such as a
monk with Empty Body using Deflect Missiles to reduce damage before
the resistance is factored in, so that the damage reduction must be
calculated and then resistance must be applied). It's also nice to be able
to confirm that the luck of a particular roll was unusually good or bad.