Add collapsible dropdowns for dice rolls
Since hover text is not readily visible and not accessible on
all devices, this adds dropdown panels with dice-roll details for
all attack and damage rolls. The hover text is retained as well,
since some users may prefer it.

This also fixes a couple of test errors introduced by a previous
commit.