Dice dropdown and formatting fixes (#177)
Since hover text is not readily visible and not accessible on
all devices, this adds dropdown panels with dice-roll details for
all attack and damage rolls. The hover text is retained as well,
since some users may prefer it.

This also makes several formatting fixes, such as adding a
default checked value to the advantage radio button. 
It limits the maximum width of the paragraph so that it remains
easily readable as the screen widens. It also limits the maximum
width of the form overall, so that it remains easier to use on
very wide screens.