Add injected service to control randomness for die rolls
The current acceptance tests for the form do not test attack details
well, since the randomness of the die rolls makes it impossible to
predict the exact format of the attack details. To fix this, this
commit introduces an injected service which manages the randomness
of die rolls. This service can be overridden in tests to force dice
to roll their maximum or minimum values (or a value in between).

Unfortunately, services can only be injected into other services or
into components. Since the component which creates dice is the
repeated attack form, and most of the abstraction layers (including
the die class) are not components or services, the randomness
service must be injected into the repeated attack form and then
passed all the way down to the dice rolled for attack rolls or
damage rolls.