Test attack form with fixed randomness behavior
This replaces the source of randomness for dice with hardcoded
behavior so that the attack rolls and damage rolls can be
predicted. This allows the test to trigger a set of attacks and
predict the resulting detail display. Unlike the integration
test for the detail-display component, this test uses data
created by the rest of the application (with predictable die
behavior) rather than hardcoded input data to be rendered. This
will catch cases where the data being sent from the backend no
longer matches the data expected by the detail-display
rendering.

The UI uses Bootstrap's "collapse" component to show details of
rolls. Unfortunately, this means that Bootstrap's Javascript
manages the collapse state of these fields. Something about this
is incompatible with the testing framework. State persists between
tests, such that if any test after the first attempts to close a
collapsible pane, the pane will not close. For now, rather than
re-implementing collapsible sections and removing Bootstrap's
state management, the tests do not check the ability to close
panels after they have been opened.