Remove link in damage display if there were no dice rolled
The damage total is typically formatted as a link, which if clicked
controls a collapsible pane displaying the actual dice rolls.
However, if there were no dice to roll, because the damage was a
constant (for instance, a marut's hit always deals 60 force damage),
then this pane holds no information. This removes the link
formatting if the collapsible pane is empty.

There is no corresponding formatting for the attack rolls, since
the attack always involves a 1d20 at minimum. (Technically, some
monsters--including the marut--do not need to roll a 1d20 and
are guaranteed to hit with every attack; but this is a rare case,
and if all attacks are guaranteed to hit then this tool is likely
not necessary.)