Extract rolling a d20 with advantage/disadvantage to new class
Adding saving throws to this tool will be a useful feature; it
will handle the case when a player character uses fireball or a
similar spell against a group of attackers. To prepare for this,
this commit extracts rolling a d20 with advantage/disadvantage
into a separate class, rather than making this behavior part of
attacks. This will allow the saving-throw class to reuse this
logic.