Add support for repeated saving throws, though not damage
This adds a class which simulates repeated saving throws, taking
modifiers and advantage/disadvantage into account, and tracks how
many of the throws passed the DC for the save. This class does not
yet handle any damage inflicted by the save to keep changes easier
to review.

Because the details for repeated attacks and the details for
repeated saves are very similar, this commit also extracts a
component to represent a display of attack/save details. Custom
display, such as "Target AC" versus "Save DC", is handled by
having each component which uses the display panel implement
functions and pass them into the component to create labels. The
repeated attack form and the repeated save form both use this
class to display the results of sets of attacks or saves.