Add repeated saving throws (#211)
This adds backend support for repeated saving throws. These will be
valuable if player characters attack a group of enemies with 
area-of-effect spells, or if a horde of enemies attack a player character
with something requiring saves rather than attack rolls. 

This commit refactors portions of the attack class to extract logic, such
as rolling a d20 with advantage/disadvantage and adding modifiers
to the result, which is shared between attacks and saving throws. 

This commit does not expose repeated saves in the user interface, nor
does it handle damage calculations related to passed or failed saves.
These features will be handled in later commits.