Add tests for repeated-save form and simplify other tests
This adds tests for the form which controls repeated saves, based
heavily on the existing tests for the attack form. It also uses
the "!" notation rather than "?" in tests where a value is
expected to be set, simplifying the tests as a result.

This also adds parent element tags to the attack and save tests.
Since the details pane is used by both the attack form and the
save form, looking up elements in tests can fail because both forms
use the same tags for elements. This adds a "#nav-attacks" and
"#nav-saves" prefix to element lookups in tests so that all of the
attack-form tests retrieve elements from the attack form and all of
the save-form tests retrieve elements from the save form. Although
some of the element labels are actually unique, all labels are
prefixed so that future work which introduces new labels will not
accidentally break the tests.