Implement damage for saving throws
This implements support for linking damage to saving throws,
including half-damage on successful saves. Currently, damage is
rolled only once and applied to all saves in the set.

To support inflicting half damage on successful saves, and only
afterwards applying resistance and vulnerability (as is required
by the definition of resistance and vulnerability), this refactors
damage handling to separate rolling damage and calculating the
damage that results from the die rolls. The total of a damage's
dice and modifiers, and the amount of damage inflicted, are
tracked separately.