Format handlebars files with prettier
This formats the handlebars files using "prettier" by adding the
"Prettier - code formatter" extension to VSCode and applying it
to all handlebars files. This causes no visible changes, but will
make the files significantly easier to work with in the future.

This also deletes a misleading "aria-describedBy" attribute on
the repeated save and attack forms since the description
referenced has been removed.