Update version of github actions and installed node
This updates the version of github actions being used to a
version which no longer uses the deprecated node 16, and
installs an updated version of node in the actions.

It also increases the delay between switching tabs and
analyzing the visible page in the test for the main page,
which is flaky when the delay is only 100 milliseconds because
the tab switch may not complete before the visibility of
elements is checked.