Indicate halved damage rolls in save detail display
This updates the save damage display to include '(halved)' when
damage has been halved by a pass in a save-for-half-damage situation.
Resistance and vulnerability are still displayed as well, and this
adds tests for both attacks and saves to verify that the display is
as expected (using fixed dice).

This also updates the fixed-dice save tests to switch to the save
tab before running the main tests. This makes debugging the tests
easier, as the save tab is displayed in the testing pane while the
tests are running.