Switch to HTML inputs for repeated attack form
This switches the repeated-attack form to use standard HTML inputs
rather than Ember Input components, as a new lint rule blocks
Inputs because they use two-way binding to modify values (which
could lead to unexpected changes in the values for classes). It
also removes an on-click function call from the submit button of
the form.