Display d20 rolls with advantage and disadvantage for attacks/saves
Although all damage dice were displayed for attacks and saves, only
one d20 roll was displayed on the UI even when the attack or save
had been made with advantage or disadvantage. This was confusing,
making it appear that the advantage or disadvantage might not be
handled correctly. This commit updates the display to show a single
d20 roll when a straight roll was made and two d20 rolls when a roll
was made with advantage or disadvantage.