Change default for passed save to half-damage
Although there are many spells and magical effects
that inflict no damage on a passed save, many of the
most popular area-of-effect spells inflict half
damage on a passed save. The default damage, for instance,
is set up for (third-level) Fireball, which inflicts
half damage on a pass. This switches the default
behavior when a save is passed to half damage.