Support evasion for targets of saves
Some D&D classes, notably monks and rogues, have the evasion
feature: when they have to make a dexterity save against a
spell which deals half damage on a successful save and full
damage on a failed save, they instead take no damage on a
failed save and half damage on a failed save. (Other things,
such as the Mounted Combatant feat, can provide similar
behavior.)

This adds a checkbox to indicate that every save is being
rolled by a target which has evasion. This checkbox is disabled
unless the spell deals half damage on a successful save, in
order to partially enforce that evasion is applied as written
in the D&D 5e player's handbook.